Extract error handling from middleware
This allows handlers to use error handling functions
directly if they want to.
